+++ src/components/ui/hero.tsx
@@ -33,57 +33,17 @@ function HeroBackground({
   return <Comp className={cn("absolute inset-0 -z-10 size-full", className)} {...props} />;
 }
 
-export function HeroBackgroundGrid({
-  gridSize = 100,
-  className,
-  ...props
-}: Omit<React.ComponentProps<"svg">, "children"> & { gridSize?: number }) {
-  return (
-    <svg
-      {...props}
-      aria-hidden="true"
-      className={cn(
-        "absolute inset-0 -z-10 size-full mask-[radial-gradient(100%_100%_at_top_right,white,transparent)]",
-        "stroke-border",
-        className
-      )}
-    >
-      <defs>
-        <pattern
-          x="50%"
-          y={-1}
-          id="983e3e4c-de6d-4c3f-8d64-b9761d1534cc"
-          width={gridSize}
-          height={gridSize}
-          patternUnits="userSpaceOnUse"
-        >
-          <path d={`M.5 ${gridSize}V.5H${gridSize}`} fill="none" />
-        </pattern>
-      </defs>
-      <svg x="50%" y={-1} className="fill-primary/10 overflow-visible">
-        <path
-          d={`M-${gridSize} 0h${gridSize + 1}v${gridSize + 1}h-${gridSize + 1}Z M${3 * gridSize} 0h${gridSize + 1}v${gridSize + 1}h-${gridSize + 1}Z M-${2 * gridSize} ${3 * gridSize}h${gridSize + 1}v${gridSize + 1}h-${gridSize + 1}Z M${gridSize} ${4 * gridSize}h${gridSize + 1}v${gridSize + 1}h-${gridSize + 1}Z`}
-          strokeWidth={0}
-        />
-      </svg>
-      <rect
-        fill="url(#983e3e4c-de6d-4c3f-8d64-b9761d1534cc)"
-        width="100%"
-        height="100%"
-        strokeWidth={0}
-      />
-    </svg>
-  );
-}
-
-function HeroContent({ size = "default", className, asChild = false, ...props }: ContainerProps) {
+function HeroContent({ className, asChild = false, ...props }: ContainerProps) {
   const Comp = asChild ? Slot : Container;
 
   return (
     <Comp
-      size={size}
-      className={cn(containerVariants({ size }), "relative z-10 py-12 sm:py-20", className)}
       {...props}
+      className={cn(
+        containerVariants({ size: props.size }),
+        "relative z-10 py-12 sm:py-20",
+        className
+      )}
     />
   );
 }